salad dressing added

+++ src/main/java/ua/training/chef/model/vegetable/Cucumber.java
@@ -1,9 +1,10 @@
 package ua.training.chef.model.vegetable;
 
+import java.math.BigDecimal;
+
 public class Cucumber extends Vegetable {
 
-	public Cucumber(String type, double caloric, double price) {
+	public Cucumber(String type, double caloric, BigDecimal price) {
 		super(type, caloric, price);
 	}
-
 }
