+++ src/main/java/ua/training/chef/model/vegetable/Tomato.java
@@ -1,9 +1,10 @@
 package ua.training.chef.model.vegetable;
 
+import java.math.BigDecimal;
+
 public class Tomato extends Vegetable {
 
-	public Tomato(String type, double caloric, double price) {
+	public Tomato(String type, double caloric, BigDecimal price) {
 		super(type, caloric, price);
 	}
-
 }
